# go: reduce memory footprint of writer