@wkalt 2024-04-26
@james-rms the measurements you show are nearly identical. I don't doubt the effect but those examples don't really show it.. does this make a big difference in the 200+ files case you mention?