@james-rms 2024-04-29
No change to concurrency behavior -the writer remains non-thread-safe and should be locked externally if people want to use it across multiple goroutines.